Add files via upload
In Lab 3, we will explore the following: 
Add output capability to the R-Pi through exploration of a number of attached components. 
Build a robot platform for the RPi 
Explore robot applications for the RPi and integrate these with piTFT feedback 
One overall goal will be the development of modular code blocks that can be reused in successive steps of the lab.
Consider software design before beginning code in order to map operations into python functions that may be used in future parts of the lab.
Plan to show design of software modules, used to achieve some degree of modularity, in your lab report 

The goal of Lab 3 will be to complete an autonomous robot under control of an application to maneuver within the environment and to display status and control information.  The robot will be self-contained and not be connected to any external devices